:fire: Refactor bot commands and handlers; implement dynamic financial assistant prompt with revenue types
