+++ src/bot/commands/__init__.py
@@ -1,0 +1,5 @@
+from .login import cmd_login
+from .start import cmd_start
+from .help import cmd_help
+from .revenue_type import cmd_revenue_types
+from .crop_variety import cmd_crop_varieties
