+++ src/bot/commands/start.py
@@ -1,0 +1,4 @@
+from aiogram.types import Message
+
+async def cmd_start(message: Message):
+    await message.answer("👋 ¡Bienvenido a LiteFarmBot! Usa /help para ver los comandos disponibles.")
