+++ src/bot/handlers/__init__.py
@@ -1,0 +1,4 @@
+from .regular_message import handle_regular_message
+from .login import handle_login_flow
+# from .revenue_types import cmd_revenue_types
+# from .crop_varieties import cmd_crop_varieties
